# Increase general compression efficiency