@pierrec 2016-12-05
Hello,

Thank you for putting efforts into this, I had this investigation on my TODO list forever.

I have no objection on removing concurrent compression on the writer since it actually helps in making it (much) faster (a quick and dirty test on a 1Gb file showed no impact.)

I would like to see a bound check in getUint32() though, as the very first step (before the unsafe use):
_ = b[3]

as this is done in most of Go's binary functions.

Thank you!