@connor4312 2016-12-06
Failure in the last commit seems unrelated, the race detection just makes things slow and resulted in Go 1.4 timing out 😛