@pierrec 2016-12-06
Sorry I wasnt clear:
_=b[3] should be the first statement in the function, so that the unsafe access is checked (as you have provided), but also so that the non little endian platforms benefit from the bound check eliminations.
I will double check the travis errors when I get time.